wip: initial port of playlist snapshot logic into worker
